Move in-memory backend into its own file
