+++ src/cli/index.ts
@@ -2,6 +2,7 @@
 import React from "react";
 import { render } from "ink";
 import meow from "meow";
+import { backend } from "../NavigatorInMemoryBackend";
 
 import App from "./App";
 
@@ -16,6 +17,8 @@ const cli = meow(`
 
 render(
   React.createElement(App, {
+    // TODO: replace with Git backend
+    backend,
     repoPath: cli.input[0] ?? process.cwd(),
   }),
 );
